Fix access ressource fct

+++ index.js
@@ -318,7 +318,6 @@ var onmessage = function(payload) {
             var token = new Token();
             token.new(request.action,3600);
             transaction_request.new(request.requested,request.requester,request.action,token);  
-            console.log(transaction_request.show());
             tmp = {
                 Transaction : transaction_request,
                 nb_node : get_nb_miner(fileAdresses),
@@ -340,7 +339,7 @@ var onmessage = function(payload) {
             var jsonTmp = JSON.stringify(objTmp);
             fs.writeFileSync(fileTmp, jsonTmp, 'utf8');
             console.log('Has access, broadcast to miner to validate transaction : '+transaction_request.hash);
-            broadcast_transaction(fileAdresses,transaction_request,'request',get_publicKey_node(fileConfig));
+            broadcast_transaction(fileAdresses,transaction_request,'request',get_publicKey_node(fileConfig),fileConfig);
             
         }
     }
@@ -495,11 +494,11 @@ var onmessage = function(payload) {
     if(message.type == 10){
         transaction=message.transaction;
 
-        var fileAccess = __dirname+'/tmp/node1/list.json';
-        var fileData = __dirname+'/tmp/node1/blocs/data.json';
-        var fileAdresses = __dirname+'/tmp/node1/adresses.json';
-        var fileConfig = __dirname+'/tmp/node1/config.json';
-        var fileTmp = __dirname+'/tmp/node1/tmp.json';
+        var fileAccess = __dirname+'/tmp/node/list.json';
+        var fileData = __dirname+'/tmp/node/blocs/data.json';
+        var fileAdresses = __dirname+'/tmp/node/adresses.json';
+        var fileConfig = __dirname+'/tmp/node/config.json';
+        var fileTmp = __dirname+'/tmp/node/tmp.json';
         var dataTmp=fs.readFileSync(fileTmp, 'utf8');
         var tmp = {
                 Transaction : transaction,
@@ -513,8 +512,7 @@ var onmessage = function(payload) {
             };
         }
         objTmp.table.push(tmp); 
-        console.log('Transaction received : ');
-        console.log(transaction);
+        console.log('Transaction received : '+transaction.hash);
         var jsonTmp = JSON.stringify(objTmp);
         fs.writeFileSync(fileTmp, jsonTmp, 'utf8');
     
@@ -523,14 +521,14 @@ var onmessage = function(payload) {
         else boolTransactionValid = verify_transaction_use(transaction,fileAccess,fileAdresses,fileData);
             
             console.log('Transaction validity : '+boolTransactionValid+' Know send response to miner');
-
+            nodeInfo=get_node_info(fileConfig);
             var packet = {
                     from: {
-                        address: server.host,
-                        port: server.port ,
+                        address: nodeInfo.Server.IP,
+                        port: nodeInfo.Server.port,
                         id: server.id
                     },
-                message: { type: 12, host: server.host, port: server.port, publicKey: get_publicKey_node(fileConfig), transactionHash : transaction.hash, typeTransaction : message.typeTransaction, response : boolTransactionValid  } 
+                message: { type: 12, host: nodeInfo.Server.IP, port: nodeInfo.Server.port, publicKey: get_publicKey_node(fileConfig), transactionHash : transaction.hash, typeTransaction : message.typeTransaction, response : boolTransactionValid  } 
             };
             server.sendMessage({address: message.host, port: message.port},packet);
     } 
@@ -569,15 +567,15 @@ var onmessage = function(payload) {
                 if(objTmp.table[i].Transaction.hash == hash){
                     if(message.response == true ) objTmp.table[i].nb_agree++ ;
                     if(message.response == false ) objTmp.table[i].nb_reject++ ;
-                    if(objTmp.table[i].nb_node/2 < objTmp.table[i].nb_agree){
+                    if(objTmp.table[i].nb_node/2 <= objTmp.table[i].nb_agree){
                         console.log('Transaction success ... Broadcast response to miner');
                         bool = true;
                         insert_Transaction(objTmp.table[i].Transaction,fileData);
                         console.log(objTmp.table[i].Transaction.requester+' can '+objTmp.table[i].Transaction.action+' '+objTmp.table[i].Transaction.requested);
                         // Inform the two node that the access is granted
-                        broadcast_response(fileAdresses,objTmp.table[i].Transaction.hash,get_publicKey_node(fileConfig),'valid');
+                        broadcast_response(fileAdresses,objTmp.table[i].Transaction.hash,get_publicKey_node(fileConfig),'valid',fileConfig);
                     }
-                    if(objTmp.table[i].nb_node/2 < objTmp.table[i].nb_reject){
+                    if(objTmp.table[i].nb_node/2 <= objTmp.table[i].nb_reject){
                         console.log('Transaction rejected ... Don\'t Broadcast response to miner');
                         bool = true;
                        // broadcast_response(fileAdresses,objTmp.table[i].tmp.Transaction.hash,get_publicKey_node(fileConfig),'novalid');
@@ -757,7 +755,7 @@ var transaction_request = new TransactionRequest();
             var jsonTmp = JSON.stringify(objTmp);
             fs.writeFileSync(fileTmp, jsonTmp, 'utf8');
             
-            broadcast_transaction(fileAdresses,transaction_use,'use',get_publicKey_node(fileConfig));
+            broadcast_transaction(fileAdresses,transaction_use,'use',get_publicKey_node(fileConfig),fileConfig);
         }       */
     
 var str = "message to sign";
@@ -983,7 +981,7 @@ function receiveNewNode(port){
             var jsonTmp = JSON.stringify(objTmp);
             fs.writeFileSync(fileTmp, jsonTmp, 'utf8');
             
-            broadcast_transaction(fileAdresses,transaction_use,'use',get_publicKey_node(fileConfig));
+            broadcast_transaction(fileAdresses,transaction_use,'use',get_publicKey_node(fileConfig),fileConfig);
     });
 
     app.post('/getAllNode',function(req, res){
@@ -1036,6 +1034,31 @@ function receiveNewNode(port){
         response='SUCCESS';
         broadcast_access_rights('ADD','','','','','',listAccess,fileAdresses);
         res.send(response);
+    });
+
+    app.post('/accessRessource',function(req,res){
+        console.log('Received request to access ressource');
+        
+        var fileConfig = __dirname+'/tmp/node/config.json';
+        
+        request = {
+                requester :req.body.requester,
+                requested : req.body.requested,
+                action : req.body.action,
+                conditions : '',
+                obligations : '',
+            }
+        // Send request to execute Action
+        var packet = {
+                from: {
+                    address: server.host,
+                    port: server.port ,
+                    id: server.id
+                    },
+                message: { type: 8, host: server.host, port: server.port, publicKey: get_publicKey_node(fileConfig), request : request } 
+            };
+        
+        server.sendMessage({address: 'localhost', port: 8001},packet);
     });
 
     app.listen(port, function() {
@@ -1141,6 +1164,15 @@ function configNode(ip,mac,role,port,fileConfig){
     fs.writeFileSync(fileConfig, jsonConfig, 'utf8');
 }
 
+function get_node_info(fileConfig){
+    var dataConfig=fs.readFileSync(fileConfig, 'utf8');
+    var objConfig = JSON.parse(dataConfig);
+    if(dataConfig.length != 0){
+        return objConfig.table[0];
+    }
+    return false;
+}
+
 function QueryPermission(fileAccess,requester,requested,action,conditions,obligations){
 
     var dataAccess=fs.readFileSync(fileAccess,'utf8');
@@ -1150,11 +1182,11 @@ function QueryPermission(fileAccess,requester,requested,action,conditions,obliga
     var boolAccess = false;
     if(dataAccess.length != 0 ){
         var objAccess = JSON.parse(dataAccess);
-        for(i=0;i<objAccess.table[0].Node.length;i++){
+        for(i=0;i<objAccess.table.length;i++){
             var boolNode = false;
-            if(objAccess.table[0].Node[i].adr == requester){
-                for(k=0;k<objAccess.table[0].Node[i].accesslist.length;k++){
-                    if(requested == objAccess.table[0].Node[i].accesslist[k].ressource && action == objAccess.table[0].Node[i].accesslist[k].rights) boolAccess = true;
+            if(objAccess.table[i].Node.adr == requester){
+                for(k=0;k<objAccess.table[i].Node.accesslist.length;k++){
+                    if(requested == objAccess.table[i].Node.accesslist[k].ressource && action == objAccess.table[i].Node.accesslist[k].rights) boolAccess = true;
                 }
             }
         }
@@ -1198,7 +1230,7 @@ function get_all_node(fileAdresses){
     return adresses;
 }
 
-function broadcast_transaction(fileAdresses,transaction,type,publicKey){
+function broadcast_transaction(fileAdresses,transaction,type,publicKey,fileConfig){
     var objAdresses = {
         table: []
         };
@@ -1209,15 +1241,17 @@ function broadcast_transaction(fileAdresses,transaction,type,publicKey){
         objAdresses = JSON.parse(dataAdresses);
 
         for(i=0;i<Object.keys(objAdresses.table).length;i++){
+
             //Test if node is miner
-            if(objAdresses.table[i].Node.role == 'miner'){
+            if(objAdresses.table[i].Node.role == 'miner' && objAdresses.table[i].Node.adr != publicKey){
+                nodeInfo = get_node_info(fileConfig);
                 var packet = {
                     from: {
-                        address: server.host,
-                        port: server.port ,
+                        address: nodeInfo.Server.IP,
+                        port: nodeInfo.Server.port,
                         id: server.id
                         },
-                    message: { type: 10, host: server.host, port: server.port, publicKey: publicKey, transaction : transaction, typeTransaction : type } 
+                    message: { type: 10, host: nodeInfo.Server.IP, port: nodeInfo.Server.port, publicKey: publicKey, transaction : transaction, typeTransaction : type } 
                 };
                 server.sendMessage({address: objAdresses.table[i].Node.IP, port: objAdresses.table[i].Node.port},packet);
             }
@@ -1225,7 +1259,7 @@ function broadcast_transaction(fileAdresses,transaction,type,publicKey){
     }
 }
 
-function broadcast_response(fileAdresses,transactionHash,publicKey,response){
+function broadcast_response(fileAdresses,transactionHash,publicKey,response,fileConfig){
     var objAdresses = {
         table: []
         };
@@ -1237,14 +1271,15 @@ function broadcast_response(fileAdresses,transactionHash,publicKey,response){
 
         for(i=0;i<Object.keys(objAdresses.table).length;i++){
             //Test if node is miner
-            if(objAdresses.table[i].Node.role == 'miner'){
+            if(objAdresses.table[i].Node.role == 'miner' && objAdresses.table[i].Node.adr != publicKey){
+                nodeInfo = get_node_info(fileConfig);
                 var packet = {
                     from: {
-                        address: server.host,
-                        port: server.port ,
+                        address: nodeInfo.Server.IP,
+                        port: nodeInfo.Server.port,
                         id: server.id
                         },
-                    message: { type: 13, host: server.host, port: server.port, publicKey: publicKey, transactionHash : transactionHash, response : response } 
+                    message: { type: 13, host: nodeInfo.Server.IP, port: nodeInfo.Server.port, publicKey: publicKey, transactionHash : transactionHash, response : response } 
                 };
                 server.sendMessage({address: objAdresses.table[i].Node.IP, port: objAdresses.table[i].Node.port},packet);
             }
@@ -1387,42 +1422,40 @@ function existNode(nodeAdr,fileAdresses){
 
 function insert_Transaction(transaction,fileData){
     var data=fs.readFileSync(fileData, 'utf8');
-     console.log('Insert transaction'+transaction.hash);  
-    if(data.length != 0){
-        obj = JSON.parse(data);
-        block=new Block();
-        var objTree = obj.table[obj.table.length-1].Block._tree;
-        var tabTree = Object.keys(objTree).map(function(key) {
-          return [objTree[key]];
-        });
-        var newBlock = false;
-        //obj.table[obj.table.length-1].Block._tree
-        block.new(obj.table[obj.table.length-1].Block.hash,obj.table[obj.table.length-1].Block.previousHash,obj.table[obj.table.length-1].Block.timestamp,obj.table[obj.table.length-1].Block.merkleRoot,obj.table[obj.table.length-1].Block.difficulty,obj.table[obj.table.length-1].Block.txs,obj.table[obj.table.length-1].Block.nonce,obj.table[obj.table.length-1].Block.no,tabTree,obj.table[obj.table.length-1].Block.numberMax);
-        //block.setTransactions() = obj.table[obj.table.length-1].Block;
-        if(block.getNumberOfTransactions() >= block.getNumberMax()){
-            //Generate new Block
-            miner.setPreviousBlock(block);
-            miner.generateHash();
-            blockNew = miner.getNewBlock();
+    console.log('Insert transaction : '+transaction.hash); 
+    if(existTransaction(transaction.hash,fileData) == false){
+        if(data.length != 0){
+            obj = JSON.parse(data);
+            block=new Block();
+            var objTree = obj.table[obj.table.length-1].Block._tree;
+            var tabTree = Object.keys(objTree).map(function(key) {
+              return [objTree[key]];
+            });
+            //obj.table[obj.table.length-1].Block._tree
+            block.new(obj.table[obj.table.length-1].Block.hash,obj.table[obj.table.length-1].Block.previousHash,obj.table[obj.table.length-1].Block.timestamp,obj.table[obj.table.length-1].Block.merkleRoot,obj.table[obj.table.length-1].Block.difficulty,obj.table[obj.table.length-1].Block.txs,obj.table[obj.table.length-1].Block.nonce,obj.table[obj.table.length-1].Block.no,tabTree,obj.table[obj.table.length-1].Block.numberMax);
+            //block.setTransactions() = obj.table[obj.table.length-1].Block;
+            if(block.getNumberOfTransactions() >= block.getNumberMax()){
+                //Generate new Block
+                miner.setPreviousBlock(block);
+                miner.generateHash();
+                blockNew = miner.getNewBlock();
 
-            var tx=blockNew.getTransactions();
-            tx.push(transaction);
-            blockNew.setTransactions(tx);
-            blockNew.previousHash=block.hash;
-            obj.table[obj.table.length-1].Block=block;
-            obj.table.push({Block : blockNew});
-            newBlock = true;
-        }else{
-            var tx=block.getTransactions();
-            tx.push(transaction);
-            block.setTransactions(tx);
-            obj.table[obj.table.length-1].Block=block;
+                var tx=blockNew.getTransactions();
+                tx.push(transaction);
+                blockNew.setTransactions(tx);
+                blockNew.previousHash=block.hash;
+                obj.table[obj.table.length-1].Block=block;
+                obj.table.push({Block : blockNew});
+            }else{
+                var tx=block.getTransactions();
+                tx.push(transaction);
+                block.setTransactions(tx);
+                obj.table[obj.table.length-1].Block=block;
+            }
+            var json = JSON.stringify(obj);
+            fs.writeFileSync(fileData, json, 'utf8');
         }
-        var json = JSON.stringify(obj);
-        fs.writeFileSync(fileData, json, 'utf8');
-
-        if(newBlock == true) return blockNew;
-    }
+    }        
 }
 
 function check_efficiency(transaction,fileAdresses,fileData,fileAccess){
